YouTube_API_Playlist_Analyzer/requirements.txt
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
Django==2.2.19
djangorestframework==3.11.2
dynamic-rest==2.0.0
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.3
googleapis-common-protos==1.53.0
httplib2==0.19.0
idna==2.10
inflection==0.3.1
oauthlib==3.1.0
packaging==20.9
pickle-mixin==1.0.2
protobuf==3.15.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
